+++ java-src/keypin/MediaReader.java
@@ -10,23 +10,12 @@
 
 package keypin;
 
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.io.Writer;
 import java.util.Map;
 
-public interface ConfigIO {
+public interface MediaReader {
 
     public String getName();
 
-    public boolean canRead(String filename);
-
-    public Map<?, ?> readConfig(InputStream in) throws Exception;
-
-    public boolean canWrite(String filename);
-
-    public void writeConfig (OutputStream out, Map<?, ?> config, boolean escape) throws Exception;
-
-    public void writeConfig (Writer out, Map<?, ?> config, boolean escape) throws Exception;
+    public Map<?, ?> readConfig(ConfigIO configReader, String resource) throws Exception;
 
 }