+++ java-src/keypin/PropertyConfigIO.java
@@ -21,6 +21,11 @@ public class PropertyConfigIO implements ConfigIO {
     public static final PropertyConfigIO INSTANCE = new PropertyConfigIO();
 
     @Override
+    public String getName() {
+        return "Properties";
+    }
+
+    @Override
     public boolean canRead(String filename) {
         return null != filename && filename.toLowerCase().endsWith(".properties");
     }